overlays: colibri-imx8x: use mxsfb driver for LCD/VGA interfaces
This driver correctly supports display flags and fixes the pixel
shift issue on displays with the data latching on the rising edge.

The DPU-enabled driver for LCDIF device (display_lcdif) doesn't
support display special flags. It prevents the VGA port on the
Colibri Evaluation board working correctly.

The driver mxsfb (adma_lcdif) works well with display flags,
so that it should be used as a default driver for LCD panels and
VGA monitors.

Note, the mxsfb driver DOESN'T support imx dpu unit.

Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@toradex.com>