verdin-imx8mp: add/modify overlays for supporting ov5640 on mezzanine
verdin-imx8mp_mezzanine-ov5640_overlay.dts:
Modify the regular overlay to work with the default jumpering of the
Mezzanine which uses the same GPIOs as the first camera interface CSI_1
and the same I2C bus as used for DSI display adapters! So, unless we
modify the OV5640 driver to cope with using them same GPIOs twice or
ignore absence of GPIOs this will not work for any concurrent dual
camera use case.

verdin-imx8mp_mezzanine-ov5640-2_overlay.dts:
Add a new overlay to work with an alternate jumpering of the Mezzanine
which uses the same Verdin GPIO2 as used for the Atmel MXT Touchscreen
and the same I2C bus as used for DSI display adapters! So, while this
can not concurrently used with our 10 inch capacitive LVDS touch screen
it does work fine for any concurrent dual camera use case.

While at it indicate GPIO usage of primary CSI_1 OV5640 overlay.

Related-to: ELB-4035
Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>